scam-honeypot-AI-main/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
groq
google-generativeai